pyproject.toml
[project]
name = "wise-parakeet"
version = "0.1.0"
description = "Basic web app for spam email classification."
authors = [
    {name = "Noah Grattan"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
